chore: 관심분야 뱃지 별도 컴포넌트로 분리
