+++ client/src/common/InterestBadge/constants.ts
